src/requirements.txt
fastapi[all]==0.86.0
uvicorn[standard]==0.18.3
pydantic>=1.8.0,<2.0.0 
pandas==1.3.5
scikit-learn<=1.0.2
joblib==1.2.0
python-dotenv
boto3 
sqlmodel==0.0.8
pymysql==1.0.2
SQLAlchemy==1.4.18
scipy==1.10.1